fix(backend): soporta cantidad negativa en API de ventas sin corromper agregados (#82)
VentaHistorica.Cantidad (uint), VentaAgregada.TotalCantidad (uint) y los
campos de VentaSkuResumen (ulong) no podian representar una nota de
credito real tras la migracion de esquema de #80 -- el cast (uint)/(ulong)
de un valor negativo en C# hace wraparound silencioso (ej. -5 -> numero
gigante), no tira excepcion. Se migran a int/long en modelos, DTOs de
respuesta e interfaces de servicio/repositorio.

Se sacan los filtros "Cantidad > 0" en VentasService.cs (TopSkusByVentas,
GetSkuResumen) que sumaban antes de agregar -- silenciosamente excluian
las devoluciones, mostrando bruto en vez del neto real que pidio el
cliente en #61/#79. Se preserva el filtro en el unico lugar que
corresponde (deteccion de "primera venta efectiva", igual criterio que
ioworker/data.py en #81).

GetAbcClassification (ResultadosService.cs) redisena el algoritmo de
Pareto: granTotal ahora solo suma SKUs de Total positivo (antes incluia
negativos, podia hacer que el acumulado superara 100% antes de terminar
de recorrer la lista, rompiendo la clasificacion de todos los SKUs
siguientes). SKUs con Total <= 0 se clasifican "C" automaticamente sin
participar del acumulado.

Hallazgo adicional durante la implementacion: VentasMensuales.VentasCantidad
(ulong) tambien mapea a ventas_mensuales.ventas_cantidad (migrada por #80)
y SI se expone via GET /api/ventasmensuales (VentasMensualesController
consulta el DbSet directo, bypaseando el repositorio) -- corregido tambien.

Verificado: build limpio, cero regresion contra datos reales locales
(ABC identico antes/despues), y un caso sintetico de neto negativo
(-30) reflejado correctamente en ABC/top-skus/sku-resumen sin clamp ni
wraparound.

+++ apps/backend/WebApi/DataAccess/Repositories/VentasDataAccess/VentaRepository.cs
@@ -77,7 +77,7 @@ namespace DataAccess.Repositories.VentaDataAccess
               Month = 0,
               Day = 0,
               Sku = g.Key.Sku,
-              TotalCantidad = (uint)g.Sum(x => (long)x.Cantidad),
+              TotalCantidad = g.Sum(x => (long)x.Cantidad),
               Quarter = g.Key.Quarter
             })
             .OrderByDescending(x => x.Year)
@@ -111,7 +111,7 @@ namespace DataAccess.Repositories.VentaDataAccess
               Month = g.Key.Month,
               Day = 1,
               Sku = g.Key.Sku,
-              TotalCantidad = (uint)g.Sum(x => x.Cantidad),
+              TotalCantidad = g.Sum(x => (long)x.Cantidad),
               Quarter = 0
             })
         : lowerPeriodo == "anual"
@@ -122,7 +122,7 @@ namespace DataAccess.Repositories.VentaDataAccess
                 Month = 0,
                 Day = 0,
                 Sku = g.Key.Sku,
-                TotalCantidad = (uint)g.Sum(x => x.Cantidad),
+                TotalCantidad = g.Sum(x => (long)x.Cantidad),
                 Quarter = 0
               })
           : q.GroupBy(v => new { v.Sku, Fecha = v.Fecha })
@@ -132,7 +132,7 @@ namespace DataAccess.Repositories.VentaDataAccess
                 Month = g.Key.Fecha.Month,
                 Day = g.Key.Fecha.Day,
                 Sku = g.Key.Sku,
-                TotalCantidad = (uint)g.Sum(x => x.Cantidad),
+                TotalCantidad = g.Sum(x => (long)x.Cantidad),
                 Quarter = 0
               });
